+++ src/dlim_api/blueprints/__init__.py
@@ -1,3 +1,3 @@
-from .sam import sam_blueprint
+from .sam_bp import sam_blueprint
 
 blueprints = [sam_blueprint]
